Enhance WebSocket error handling and logging for improved debugging

+++ cmd/main.go
@@ -18,6 +18,7 @@ func (wsh WebSocketService) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	conn, err := wsh.handler.Upgrader.Upgrade(w, r, nil)
 	if err != nil {
 		log.Println("Upgrade Error:", err)
+		return
 	}
 
 	defer conn.Close()
@@ -32,30 +33,42 @@ func (wsh WebSocketService) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 				newPublishService.RemoveConnection(conn)
 				return
 			}
+			if websocket.IsUnexpectedCloseError(err, websocket.CloseGoingAway, websocket.CloseAbnormalClosure) {
+				log.Printf("Unexpected close error: %v\n", err)
+				newPublishService.RemoveConnection(conn)
+				return // Add return here to prevent further execution with a broken connection
+			}
 			log.Println("Read Error:", err)
-			newPublishService.RemoveConnection(conn)
+			return
 		}
 		err = json.Unmarshal(msg, &message)
 		if err != nil {
-			log.Println("JSON Unmarshal Error:", err)
+			log.Printf("JSON Unmarshal Error: %v in message: %s", err, string(msg))
 			continue
 		}
 		if message.Role == "consumer" {
 			if message.Subscribe {
 				newPublishService.AddSubscribers(message.Topic, conn)
+				log.Println("Subscribed to topic:", message.Topic)
 			} else {
 				msg, is_data := newQueueService.DeQueue(message.Topic)
+
 				if is_data {
 					conn.WriteJSON(msg)
+					log.Printf("Data sent to consumer: %s\n", msg)
+				} else {
+					log.Println("No data in queue")
 				}
+
 			}
 		} else if message.Role == "producer" {
 			if message.TransmissionMode == "buffered" {
 				newQueueService.EnQueue(message.Topic, message.Message)
+				log.Printf("Message buffered: %s\n", message.Message)
 
 			} else if message.TransmissionMode == "broadcast" {
 				newPublishService.SendMessageToSubscribers(message)
-
+				log.Printf("Message broadcasted: %s\n", message.Message)
 			} else {
 				log.Printf("Invalid TransmissionMode %s\n", message.TransmissionMode)
 			}
